fix: convert timedelta to int for unit of adaptive sampling

+++ data_processor.py
@@ -127,7 +127,9 @@ class DataProcessor:
         return len(histories) - start_idx
 
     def adaptive_sampling(self, histories: pd.DataFrame) -> pd.DataFrame:
-        unit = (histories["timestamp"].iloc[-1] - histories["timestamp"].iloc[0]) / 240
+        unit = (
+            (histories["timestamp"].iloc[-1] - histories["timestamp"].iloc[0]) / 240
+        ).astype("int64")
         ts = histories["timestamp"].astype("int64").to_numpy()
 
         min_vals = np.empty(len(histories))
